requirements.txt
python-telegram-bot
pandas
requests
openpyxl